wip Switch remote state to Tanstack query
